requirements.txt
flask
Flask-SQLAlchemy
PyMySQL
requests